+++ CHANGELOG.rst
@@ -11,9 +11,22 @@ v34.5.0 (unreleased)
   datafile_resource fields do not have a value.
   https://github.com/nexB/scancode.io/issues/1177
 
+- Add a new `CollectPygmentsSymbolsAndStrings` pipeline (addon) for collecting source
+  symbol, string and comments using Pygments.
+  https://github.com/nexB/scancode.io/pull/1179
+
+- Workaround an issue with the cyclonedx-python-lib that does not allow to load
+  SBOMs that contains properties with no values.
+  https://github.com/nexB/scancode.io/issues/1185
+
+- Add a new `CollectTreeSitterSymbolsAndStrings` pipeline (addon) for collecting source
+  symbol and string using tree-sitter.
+  https://github.com/nexB/scancode.io/pull/1181
+
 - Fix `inspect_packages` pipeline to properly link discovered packages and dependencies to
   codebase resources of package manifests where they were found. Also correctly assign
-  the datasource_ids atrribute for packages and dependencies. 
+  the datasource_ids atrribute for packages and dependencies.
+  https://github.com/nexB/scancode.io/pull/1180/
 
 v34.4.0 (2024-04-22)
 --------------------